+++ Scripts/exec-login-rpi.sh
@@ -10,7 +10,7 @@
 #IP_RPI=172.28.107.227   #IP to network eduroam
 #IP_RPI=192.168.0.102    #IP to network RedeDrone
 #IP_RPI=192.168.43.133   #IP to network redeDroneC (cell phone veronica)
-IP_RPI=192.168.43.9     #IP to network RedeDroneCjes (cell phone jesimar)
+IP_RPI=192.168.43.8     #IP to network RedeDroneCjes (cell phone jesimar)
 #IP_RPI=10.42.0.199      #IP to network RedeDronePC
 #IP_RPI=192.168.0.100    #IP to network Alice