+++ UAV-IFA/src/uav/ifa/module/security_manager/SecurityManager.java
@@ -247,6 +247,9 @@ public class SecurityManager {
                         double timeDiff = (timeActual - timeInit) / 1000.0;
                         drone.setTime(timeDiff);
                         dataAcquisition.getAllInfoSensors();
+                        if (configGlobal.hasPowerModule()){
+                            dataAcquisition.getBattery();
+                        }
                         checkStatusSystem();
                         printLogAircraft.println(drone.toString());
                         printLogAircraft.flush();