+++ UAV-SOA-Interface/views.py
@@ -227,12 +227,13 @@ def getAllInfoSensors(request):
     print 'dist-to-home: ', dist_to_home
     print 'dist-to-current-wpt (%s): %s' % (next_waypoint, dist_to_current_waypoint)
     return {
-        'all-sensors': [float('%.10g' % gps.lat), float('%.10g' % gps.lon), float('%.2g' % gps.alt), 
-        float('%.2g' % alt_abs), float('%.3g' % bat.voltage), float('%.2g' % bat.current), bat.level, 
-        float('%.4g' % att.pitch), float('%.4g' % att.yaw), float('%.4g' % att.roll), vehicle.heading, 
-        float('%.2g' % vehicle.groundspeed), float('%.2g' % vehicle.airspeed), vehicle.gps_0.fix_type, 
-        vehicle.gps_0.satellites_visible, vehicle.gps_0.eph, vehicle.gps_0.epv, vehicle.velocity, 
-        next_waypoint, count_waypoint, dist_to_home, dist_to_current_waypoint, vehicle.mode.name, 
+        'all-sensors': [float('%.10g' % gps.lat), float('%.10g' % gps.lon), 
+        float('%.2g' % gps.alt), float('%.2g' % alt_abs), float('%.4g' % att.pitch), 
+        float('%.4g' % att.yaw), float('%.4g' % att.roll), vehicle.heading, 
+        float('%.2g' % vehicle.groundspeed), float('%.2g' % vehicle.airspeed), 
+        vehicle.gps_0.fix_type, vehicle.gps_0.satellites_visible, 
+        vehicle.gps_0.eph, vehicle.gps_0.epv, vehicle.velocity, next_waypoint, 
+        count_waypoint, dist_to_home, dist_to_current_waypoint, vehicle.mode.name, 
         vehicle.system_status.state, vehicle.armed, vehicle.is_armable, vehicle.ekf_ok]
     }